Update blackjack.rb
A few issues here that are causing the file not to run. You are missing an end to close your play method. You are also calling an instance method on your Deck class, which is throwing a no method error. Just remember to take small steps and I'm sure you can fix these issues. Other than that, great job.

(Do not focus on making these corrections this weekend and we will talk more about this on Monday)